tribe: move cli directory one level up
It was initially put under ui directory to pair with the gtk
directory. But after some consideration, and seeing what
other projects do, we came to the conclusion that it is not
necessary to have it so deep and nested there in the folder
hierarchy.

When translated to English, it reads like this:
"The src/cli directory holds the source code for the command
line interface tool."

Signed-off-by: Thiago Farina <tfransosi@gmail.com>