tribe: add .in extension to config.h.cmake file
Two reasons for doing this. First, in the autotools build
system, a config.h.in file is generated by autoheader when
it scans configure.ac file. So, that is one main reason,
the file generated by autoheader has .in (for input) extension.
Second, the documentation for the AC_CONFIG_HEADERS macro says
"usually the input file is named header.in".

So, let's have our version with .in extension too, to keep
moving forward on reducing the differences between the two
build systems as this will make porting to autotools build
sytem easier.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>