tribe: rename addrbook_window_new() to create_main_window
This addrbook_window_new() got its name from "sab_window_new"
when I did commit 8ca7e90d6a3c ("sab: rename 'sab_window.[ch]'
module to just 'window.[ch]' to remove the 'sab' prefix"). At
that time, the prefix "addrbook" seemed good to not leave the
function name as generic as "window_new".

The Glib documentation says "Use prefixing to avoid namespace
conflicts with other projects. If your library (or application)
is named Viewer, prefix all your function names with viewer_.
For example: viewer_file_save.". [1]

The proble is that, although it was trying to follow a naming
convention (I didn't know it was glib at that time), it is not
getting it right. "addrbook" is not the name of the application,
not an abbreviation of it, so it should not be used as a prefix
for our function names.

So, while we are not fully following the glib naming convention,
let's rename it to create_main_window, which is
verb + adjectival noun.

Refs:
[1] - https://docs.gtk.org/gobject/concepts.html#conventions

Signed-off-by: Thiago Farina <tfransosi@gmail.com>