tribe: bump gtk version to 2.16.1
In commit 37081ed13759 ("tribe: force pkg-config to check for
a minimum specific version of gtk+-2.0") I specified 2.10.0
as the minimum version for gtk. I took that version from
xarchiver but I don't remember the reason for this.
I looked at https://gitlab.gnome.org/GNOME/gtk/-/tags?sort=version_asc,
and the oldest version available with a tag is 2.16.1. This
version was released 15 years ago on April 11, 2009. We are
very late in the game to catch up with them.

The version I have on my system is 2.24.30:
> pkg-config --modversion gtk+-2.0
2.24.30

Which was tagged and released 9 years ago on Marc 5, 2016,
https://gitlab.gnome.org/GNOME/gtk/-/tags/2.24.30.

So, let's move on to 2.16.1 first and consider moving to
2.24.30 next time.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>