amhello: bump gtk version to 2.16.1
When I did commit ad849ae5ad80 ("c/apps/gui/gtk: add an example
of how to create a GTK project using GNU Autotools") I specified
the minimum version of gtk to 2.10.0, because that was the version
tribe was using at that time, which was done at commit
37081ed13759 ("tribe: force pkg-config to check for a minimum
specific version of gtk+-2.0").

We bumped the version in tribe at commit a0ad7d1b514b ("tribe:
bump gtk version to 2.16.1")

So, let's keep them in sync and move on to 2.16.1.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>