c: add Makefile.am to build sqlite3 library with autotools
This is not very ideal, because nothing there has been ported
to autotools build system, but I don't want to forget how
to build a static library in automake. So, let's have it for now.

See https://www.gnu.org/software/automake/manual/html_node/A-Library.html,
for more information on noinst_LIBRARIES.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>