hello: call configure via $srcdir for location independence
Update the bootstrap script to invoke configure using $srcdir
instead of ./ so that it runs correctly even when autogen.sh is
executed from outside the source tree.

Since the script now returns to the original working directory
($ORIGDIR) after running autoreconf, configure is no longer
located in the current directory. Calling it as:
"$srcdir/configure" "$@"

ensures the correct script is executed regardless of where the user
runs autogen.sh, matching the behavior of established GNU and GNOME
build systems.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>