echod: harden __GNUC__ feature check in optimization.h
Updates the conditional enabling __builtin_expect() to use a safer test:
    #if defined(__GNUC__) && (__GNUC__ >= 3)

This avoids evaluating __GNUC__ when it is not defined and makes the intent
explicit: only compilers that define __GNUC__ and support GCC 3+ semantics
will enable the branch-prediction macros.

No functional changes to M_LIKELY/M_UNLIKELY behavior are introduced.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>